Move all scalacheck tests to NewTreeCheck
And ignore them
